image-emedding-model/requirements.txt
Flask
Pillow
torchvision
